+++ api/app/routers/actionlogs.py
@@ -5,7 +5,7 @@ from fastapi import APIRouter, Depends, HTTPException, status
 from sqlalchemy.orm import Session
 
 from app import models, persistence, schemas
-from app.account import get_current_user
+from app.auth.account import get_current_user
 from app.database import get_db
 from app.routers.validators.account_validator import check_pteam_membership
