+++ api/app/routers/actions.py
@@ -6,7 +6,7 @@ from fastapi.responses import Response
 from sqlalchemy.orm import Session
 
 from app import models, persistence, schemas
-from app.account import get_current_user
+from app.auth.account import get_current_user
 from app.business.tag_business import check_topic_action_tags_integrity
 from app.database import get_db
 from app.detector.vulnerability_detector import fix_threats_for_topic
