+++ api/app/routers/tags.py
@@ -6,7 +6,7 @@ from fastapi.responses import Response
 from sqlalchemy.orm import Session
 
 from app import command, models, persistence, schemas
-from app.account import get_current_user
+from app.auth.account import get_current_user
 from app.business.tag_business import get_or_create_topic_tag
 from app.database import get_db
