+++ api/app/routers/threat.py
@@ -4,7 +4,7 @@ from fastapi import APIRouter, Depends, HTTPException, Query, status
 from sqlalchemy.orm import Session
 
 from app import command, models, persistence, schemas
-from app.account import get_current_user
+from app.auth.account import get_current_user
 from app.database import get_db
 from app.routers.validators.account_validator import check_pteam_membership
